refactor: refactor types, remove unused cli type
